refactor attribute
