+++ tamr_unify_client/attribute/type.py
@@ -1,13 +1,14 @@
-from tamr_unify_client.models.attribute.subattribute import SubAttribute
+from tamr_unify_client.attribute.subattribute import SubAttribute
 
 
 class AttributeType:
     """
-    The type of an :class:`~tamr_unify_client.models.attribute.resource.Attribute` or :class:`~tamr_unify_client.models.attribute.subattribute.SubAttribute`.
+    The type of an :class:`~tamr_unify_client.attribute.resource.Attribute` or :class:`~tamr_unify_client.attribute.subattribute.SubAttribute`.
+
     See https://docs.tamr.com/reference#attribute-types
 
     :param data: JSON data representing this type
-    :type: :py:class:`dict`
+    :type data: :py:class:`dict`
     """
 
     def __init__(self, data):
@@ -20,7 +21,7 @@ class AttributeType:
 
     @property
     def inner_type(self):
-        """:type: :class:`~tamr_unify_client.models.attribute.type.AttributeType`"""
+        """:type: :class:`~tamr_unify_client.attribute.type.AttributeType`"""
         if "innerType" in self._data:
             return AttributeType(self._data.get("innerType"))
         else:
@@ -28,7 +29,7 @@ class AttributeType:
 
     @property
     def attributes(self):
-        """:type: list[:class:`~tamr_unify_client.models.attribute.subattribute.SubAttribute`]"""
+        """:type: list[:class:`~tamr_unify_client.attribute.subattribute.SubAttribute`]"""
         collection_json = self._data.get("attributes")
         return [SubAttribute(attr) for attr in collection_json]
