+++ tamr_unify_client/models/dataset/resource.py
@@ -1,7 +1,7 @@
 import simplejson as json
 
+from tamr_unify_client.attribute.collection import AttributeCollection
 from tamr_unify_client.base_resource import BaseResource
-from tamr_unify_client.models.attribute.collection import AttributeCollection
 from tamr_unify_client.models.dataset.profile import DatasetProfile
 from tamr_unify_client.models.dataset.status import DatasetStatus
 from tamr_unify_client.models.dataset.usage import DatasetUsage
@@ -50,7 +50,7 @@ class Dataset(BaseResource):
         """Attributes of this dataset.
 
         :return: Attributes of this dataset.
-        :rtype: :class:`~tamr_unify_client.models.attribute.collection.AttributeCollection`
+        :rtype: :class:`~tamr_unify_client.attribute.collection.AttributeCollection`
         """
         alias = self.api_path + "/attributes"
         return AttributeCollection(self.client, alias)
